requirements.txt
# GUI
PyQt5==5.15.9
PyQt5-tools==5.15.9.3.3  # Sửa thành phiên bản mới nhất

# Core Dependencies
PyJWT==2.8.0
cryptography==41.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.0
PyYAML==6.0.0

# Protocol Dependencies
pymodbus==3.5.0
asyncua==1.0.0  # For OPC UA
paho-mqtt==2.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-qt==4.2.0

# Utilities
pycryptodome==3.19.0